Refactor tests
* separate batch, extensions and the regular apis into separate
* added delete for some tests
* drop '_' for is_k8s_running
* remove comments that do not add value
